Split tests in isolated modules
