+++ tests/fixtures/secure_api/swagger.yaml
