+++ tests/fixtures/secure_endpoint/swagger.yaml
@@ -1,0 +1,39 @@
+swagger: "2.0"
+
+info:
+  title: "{{title}}"
+  version: "1.0"
+
+basePath: /v1.0
+
+securityDefinitions:
+    oauth:
+        type: oauth2
+        flow: password
+        tokenUrl: https://ouath.example/token
+        x-tokenInfoUrl: https://ouath.example/token_info
+        scopes:
+            myscope: can do stuff
+
+paths:
+  /byesecure/{name}:
+    get:
+      summary: Generate goodbye
+      description: Generates a goodbye message.
+      operationId: fakeapi.hello.get_bye_secure
+      security:
+        - oauth:
+            - myscope
+      produces:
+        - text/plain
+      responses:
+        200:
+          description: goodbye response
+          schema:
+            type: string
+      parameters:
+        - name: name
+          in: path
+          description: Name of the person to say bye.
+          required: true
+          type: string
